Standardizing registry for the rest of the backends
