--- ibis/backends/pyspark/operations.py
@@ -1,5 +1,0 @@
-import ibis.expr.operations as ops
-
-
-class PySparkTable(ops.DatabaseTable):
-    pass
